支付功能开发

+++ 01.项目源码/trunk/0.0.1/clothing-server/src/main/java/com/dower/sharerideapp/core/serverdb/dao/PaymentFlowMapper.java
@@ -10,7 +10,7 @@ public interface PaymentFlowMapper {
 
     int deleteByExample(PaymentFlowExample example);
 
-    int deleteByPrimaryKey(Integer numId);
+    int deleteByPrimaryKey(Long numId);
 
     int insert(PaymentFlow record);
 
@@ -18,7 +18,7 @@ public interface PaymentFlowMapper {
 
     List<PaymentFlow> selectByExample(PaymentFlowExample example);
 
-    PaymentFlow selectByPrimaryKey(Integer numId);
+    PaymentFlow selectByPrimaryKey(Long numId);
 
     int updateByExampleSelective(@Param("record") PaymentFlow record, @Param("example") PaymentFlowExample example);
